requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
scikit-surprise